Fix and Simplify Merge Logic
The merge logic for splicer considers the results
for each slice to contain the same aggregate_tags and
tag fields. Sometimes this is not the case. TSDB promotes
an aggregated_tag field to a tag field if there is only
one value for it. Because of this, our signatures are
different, and the merge fails.

Jira: TDB-206
Signed-off-by: Arjun Satish <arjun.satish@turn.com>